requirements.txt
imutils
tensorflow
numpy
streamlit
Pillow
opencv-python-headless
